add: notification resources

+++ app/src/main/java/com/acmvit/c2c2021/model/NotificationItem.kt
@@ -2,5 +2,5 @@ package com.acmvit.c2c2021.model
 
 data class NotificationItem(
     val message: String = "",
-    val time: Double = 0.0
+    val time: Long = 0
 )
